Various updates including modifying project folder
@MartinHenriksen @al-fisher @EddyDalton , modified the project folder to
match the other repos, allowing for multiple projects in the
RobotToolkit which we'll invariably need. Added  minimal functionality
to keep up to date with BHoM progress. There is still a lot of work to
change this repo to object oriented from functional, and break the file
structure down into smaller chunks. More updates to follow this week.